requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1
streamlit>=1.28.0
ipywidgets>=8.0.0
pyyaml>=6.0
tqdm>=4.65.0
numba>=0.57.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0